Backend/requirements.txt
fastapi
uvicorn
opencv-python
torch
torchvision
pillow
pygame
mediapipe
numpy